+++ src/grandorgue/yaml/GOYamlModel.cpp
@@ -5,12 +5,29 @@
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
 
-#include "GOYamlInModel.h"
+#include "GOYamlModel.h"
 
+#include <wx/datetime.h>
 #include <wx/intl.h>
 #include <wx/wfstream.h>
+#include <yaml-cpp/yaml.h>
 
 #include "GOSaveableToYaml.h"
+#include "go-wx-yaml.h"
+#include "go_defs.h"
+
+static const uint8_t UTF8_BOM[] = {0xEF, 0xBB, 0xBF};
+
+const char *const INFO = "info";
+const char *const CONTENT_TYPE = "content-type";
+const char *const ORGAN_NAME = "organ-name";
+const char *const GRANDORGUE_VERSION = "grandorgue-version";
+const char *const SAVED_TIME = "saved_time";
+
+static wxString get_info_field(
+  const YAML::Node &globalNode, const wxString &fieldName) {
+  return globalNode[INFO][fieldName].as<wxString>();
+}
 
 static std::vector<char> load_file_bytes(const wxString &filePath) {
   wxFile file;
@@ -56,12 +73,12 @@ static YAML::Node load_yaml_from_file(const wxString &fileName) {
   return YAML::Load(fileContentInUtf8.data());
 }
 
-GOYamlInModel::GOYamlInModel(
+GOYamlModel::In::In(
   const wxString &organName,
   const wxString &fileName,
   const wxString &contentType)
   : m_GlobalNode(load_yaml_from_file(fileName)) {
-  const wxString fileContentType = getFileContentType(m_GlobalNode);
+  const wxString fileContentType = get_info_field(m_GlobalNode, CONTENT_TYPE);
 
   if (fileContentType != contentType)
     throw wxString::Format(
@@ -71,8 +88,48 @@ GOYamlInModel::GOYamlInModel(
       contentType);
 }
 
-const GOYamlInModel &GOYamlInModel::operator>>(
+wxString GOYamlModel::In::GetFileOrganName() const {
+  return get_info_field(m_GlobalNode, ORGAN_NAME);
+}
+
+const GOYamlModel::In &GOYamlModel::In::operator>>(
   GOSaveableToYaml &saveableObj) const {
   m_GlobalNode >> saveableObj;
   return *this;
+}
+
+GOYamlModel::Out::Out(const wxString &organName, const wxString &contentType)
+  : m_GlobalNode(YAML::NodeType::Map) {
+  YAML::Node infoNode = m_GlobalNode[INFO];
+
+  infoNode[CONTENT_TYPE] = contentType;
+  infoNode[ORGAN_NAME] = organName;
+  infoNode[GRANDORGUE_VERSION] = APP_VERSION;
+  infoNode[SAVED_TIME] = wxDateTime::Now().Format();
+}
+
+GOYamlModel::Out &GOYamlModel::Out::operator<<(
+  const GOSaveableToYaml &saveableObj) {
+  m_GlobalNode << saveableObj;
+  return *this;
+}
+
+wxString GOYamlModel::Out::writeTo(const wxString &fileName) {
+  wxString errMsg;
+  wxFileOutputStream fOS(fileName);
+
+  if (fOS.IsOk()) {
+    YAML::Emitter outYaml;
+
+    outYaml << YAML::BeginDoc << m_GlobalNode;
+
+    if (
+      !fOS.WriteAll(UTF8_BOM, sizeof(UTF8_BOM))
+      || !fOS.WriteAll(outYaml.c_str(), outYaml.size()))
+      errMsg.Printf(
+        wxT("Unable to write all the data to the file '%s'"), fileName);
+    fOS.Close();
+  } else
+    errMsg.Printf(wxT("Unable to open the file '%s' for writing"), fileName);
+  return errMsg;
 }
